Rename spl-tokens file
