+++ tokens/src/commands.rs
@@ -1,7 +1,7 @@
 use crate::{
     args::{BalancesArgs, DistributeTokensArgs, StakeArgs, TransactionLogArgs},
     db::{self, TransactionInfo},
-    spl_token_helpers::*,
+    spl_token::*,
     token_display::Token,
 };
 use chrono::prelude::*;
