+++ tokens/src/lib.rs
@@ -2,5 +2,5 @@ pub mod arg_parser;
 pub mod args;
 pub mod commands;
 mod db;
-pub mod spl_token_helpers;
+pub mod spl_token;
 pub mod token_display;
